+++ e2e/Utils.ts
@@ -14,6 +14,7 @@ import {ExecaReturnValue, sync as spawnSync} from 'execa';
 import makeDir = require('make-dir');
 import rimraf = require('rimraf');
 import dedent = require('dedent');
+import which = require('which');
 
 interface RunResult extends ExecaReturnValue {
   status: number;
@@ -47,7 +48,7 @@ export const linkJestPackage = (packageName: string, cwd: Config.Path) => {
   const destination = path.resolve(cwd, 'node_modules/', packageName);
   makeDir.sync(destination);
   rimraf.sync(destination);
-  fs.symlinkSync(packagePath, destination, 'dir');
+  fs.symlinkSync(packagePath, destination, 'junction');
 };
 
 export const makeTemplate = (
@@ -105,6 +106,7 @@ export const writeSymlinks = (
     fs.symlinkSync(
       path.resolve(directory, ...fileOrPath.split('/')),
       path.resolve(directory, ...symLinkPath.split('/')),
+      'junction',
     );
   });
 };
@@ -259,4 +261,20 @@ export const normalizeIcons = (str: string) => {
   return str
     .replace(new RegExp('\u00D7', 'g'), '\u2715')
     .replace(new RegExp('\u221A', 'g'), '\u2713');
+};
+
+// Certain environments (like CITGM and GH Actions) do not come with mercurial installed
+let hgIsInstalled: boolean | null = null;
+
+export const testIfHg = (...args: Parameters<typeof test>) => {
+  if (hgIsInstalled === null) {
+    hgIsInstalled = which.sync('hg', {nothrow: true}) !== null;
+  }
+
+  if (hgIsInstalled) {
+    test(...args);
+  } else {
+    console.warn('Mercurial (hg) is not installed - skipping some tests');
+    test.skip(...args);
+  }
 };